+++ modules/photons_transport/comms/result.py
@@ -9,10 +9,10 @@ class Result(asyncio.Future):
     options
     """
 
-    def __init__(self, request, did_broadcast, retry_options):
+    def __init__(self, request, did_broadcast, retry_gaps):
         self.request = request
+        self.retry_gaps = retry_gaps
         self.did_broadcast = did_broadcast
-        self.retry_options = retry_options
 
         self.results = []
         self.last_ack_received = None
@@ -80,7 +80,7 @@ class Result(asyncio.Future):
         """
         current = getattr(self, attr)
         asyncio.get_event_loop().call_later(
-            self.retry_options.finish_multi_gap, self.maybe_finish, current, attr
+            self.retry_gaps.finish_multi_gap, self.maybe_finish, current, attr
         )
 
     def maybe_finish(self, last, attr):
@@ -155,9 +155,7 @@ class Result(asyncio.Future):
             if self.results:
                 return True
 
-            return (
-                time.time() - self.last_ack_received
-            ) < self.retry_options.gap_between_ack_and_res
+            return (time.time() - self.last_ack_received) < self.retry_gaps.gap_between_ack_and_res
 
         elif self.request.ack_required and self.last_ack_received is not None:
             return True
@@ -167,9 +165,7 @@ class Result(asyncio.Future):
                 return False
 
             if self.num_results > 0:
-                return (
-                    time.time() - self.last_res_received
-                ) < self.retry_options.gap_between_results
+                return (time.time() - self.last_res_received) < self.retry_gaps.gap_between_results
 
             return True