+++ modules/photons_transport/retry_options.py
@@ -1,61 +1,94 @@
 from photons_app import helpers as hp
 
+from delfick_project.norms import dictobj, sb
 import time
 
 
-class RetryOptions:
-    """
-    Options for working out how long to wait for replies to our messages
+def Gaps(*, gap_between_results, gap_between_ack_and_res, timeouts):
+    default_timeouts = timeouts
+    default_gap_between_results = gap_between_results
+    default_gap_between_ack_and_res = gap_between_ack_and_res
 
-    finish_multi_gap
-        When a packet has an unbound number of results or acks, this number is
-        used to schedule the next check to see if we should finish this result
+    class tuple_spec(sb.Spec):
+        def setup(self, *specs):
+            self.spec = sb.tuple_spec(*specs)
 
-    gap_between_results
-        When a packet has an unbound number of results we uses this number to
-        determine when we have enough results. Essentially the answer is yes if
-        it's been this long since the last result
+        def normalise_filled(self, meta, val):
+            if isinstance(val, list):
+                val = tuple(val)
+            return self.spec.normalise(meta, val)
 
-        It is a good idea to make this number less than finish_multi_gap so that
-        when we check after a finish_multi_gap amount of time in the future we
-        can mark the result as done
+    class timeouts_default_spec(sb.Spec):
+        def normalise_empty(self, meta):
+            return default_timeouts
 
-    gap_between_ack_and_res
-        When a packet has a received an acknowledgment but not a result, this
-        number is used to determine if we should wait for the result or send
-        the request again.
+        def normalise_filled(self, meta, val):
+            return sb.listof(tuple_spec(sb.float_spec(), sb.float_spec())).normalise(meta, val)
 
-        i.e. only send a retry if it's been longer than this time since the
-        acknowledgement
+    class ResultGaps(dictobj.Spec):
+        gap_between_results = dictobj.Field(
+            sb.float_spec,
+            default=default_gap_between_results,
+            help="""
+            When a packet has an unbound number of results we uses this number to
+            determine when we have enough results. Essentially the answer is yes if
+            it's been this long since the last result
 
-    next_check_after_wait_for_result
-        If we should wait for the next reply instead of sending a retry, then
-        we use this time to schedule the next check.
+            It is a good idea to make this number less than finish_multi_gap so that
+            when we check after a finish_multi_gap amount of time in the future we
+            can mark the result as done
+            """,
+        )
 
-    timeouts
-        A list of (step, end) tuples that is used to determine the retry backoff.
+        gap_between_ack_and_res = dictobj.Field(
+            sb.float_spec,
+            default=default_gap_between_ack_and_res,
+            help="""
+            When a packet has a received an acknowledgment but not a result, this
+            number is used to determine if we should wait for the result or send
+            the request again.
 
-        Essentially, starting with the first step, increase by step until you
-        reach end and then use the next tuple to determine backoff from there.
+            i.e. only send a retry if it's been longer than this time since the
+            acknowledgement
+            """,
+        )
 
-        So ``[(0.1, 0.1), (0.2, 0.5), (0.5, 3)]`` would go
-        ``0.1, 0.3, 0.5, 1, 1.5, 2, 2.5, 3, 3, 3, ...``
-    """
+        timeouts = dictobj.Field(
+            timeouts_default_spec,
+            help="""
+            A list of (step, end) tuples that is used to determine the retry backoff.
 
-    finish_multi_gap = 0.4
-    gap_between_results = 0.35
-    gap_between_ack_and_res = 0.2
+            Essentially, starting with the first step, increase by step until you
+            reach end and then use the next tuple to determine backoff from there.
 
-    next_check_after_wait_for_result = 0.15
+            So ``[(0.1, 0.1), (0.2, 0.5), (0.5, 3)]`` would go
+            ``0.1, 0.3, 0.5, 1, 1.5, 2, 2.5, 3, 3, 3, ...``
+            """,
+        )
 
-    timeouts = [(0.2, 0.2), (0.1, 0.5), (0.2, 1), (1, 5)]
+        @property
+        def finish_multi_gap(self):
+            """
+            When a packet has an unbound number of results or acks, this number is
+            used to schedule the next check to see if we should finish this result
 
-    def __init__(self, *, timeouts=None, name=None):
+            This defaults to being ``gap_between_results + 0.05``
+            """
+            return self.gap_between_results + 0.05
+
+        def retry_ticker(self, name=None):
+            return RetryTicker(timeouts=self.timeouts, name=name)
+
+    return ResultGaps.FieldSpec()
+
+
+class RetryTicker:
+    def __init__(self, *, timeouts, name=None):
         self.name = name
+        self.timeouts = timeouts
+
         self.timeout = None
         self.timeout_item = None
-        if timeouts is not None:
-            self.timeouts = timeouts
 
     async def tick(self, final_future, timeout, min_wait=0.1):
         timeouts = list(self.timeouts)
@@ -65,7 +98,7 @@ class RetryOptions:
             final_future=final_future,
             max_time=timeout,
             min_wait=min_wait,
-            name=f"RetryOptions({self.name})::tick[ticker]",
+            name=f"RetryTicker({self.name})::tick[ticker]",
         )
 
         start = time.time()